requirements.txt
datasets==3.4.0
fastapi==0.115.11
numpy==2.2.4
openai==1.68.0
pydantic==2.10.3
python-dotenv==1.0.1
supabase==2.13.0
uvicorn==0.34.0